念のため8bit量子化対応

+++ lib/utils/get_model.py
@@ -39,10 +39,10 @@ def get_model(model_name, accelerator=None, add_size=False, get_custom=False):
                 )
             elif "Llama" in model_name or "vicuna" in model_name:
                 model = CustomLlamaModelForCausalLM.from_pretrained(
-                    model_name, device_map='auto',torch_dtype=torch.bfloat16)
+                    model_name, device_map='auto',torch_dtype=torch.float16)
             elif "Mistral" in model_name:
                 model = CustomMistralModelForCausalLM.from_pretrained(
-                    model_name, device_map='auto',torch_dtype=torch.bfloat16)
+                    model_name, device_map='auto',torch_dtype=torch.float16)
             else:
                 raise ValueError("")
             if add_size:
@@ -55,10 +55,10 @@ def get_model(model_name, accelerator=None, add_size=False, get_custom=False):
             )
         elif "Llama" in model_name or "vicuna" in model_name:
             model = CustomLlamaModelForCausalLM.from_pretrained(
-                model_name,torch_dtype=torch.bfloat16)
+                model_name,torch_dtype=torch.float16)
         elif "Mistral" in model_name:
             model = CustomMistralModelForCausalLM.from_pretrained(
-                model_name,torch_dtype=torch.bfloat16)
+                model_name,torch_dtype=torch.float16)
         else:
             raise ValueError("")
         if add_size:
